feat(case-detail): add panel-pilot / panel-profile-vault / panel-reference
