+++ app/(root)/(case)/case-portal/mock.ts
@@ -4,66 +4,11 @@ export const mockCaseList = [
   {
     id: '44c6cd75-b7c4-4e27-b643-ab14c15ee3a0',
     title: '知识产权案例（已更新）',
-    caseName: 'Ahmed Hassan1',
     caseType: 'Skilled Worker Visa',
     description: '这是一个更新后的知识产权侵权案例描述',
     profileId: 'profile-123',
     clientId: null,
     status: CaseStatusEnum.ANALYZING,
-    startDate: null,
-    endDate: null,
-    clientName: null,
-    profileName: null,
-    createdAt: '2025-05-19T09:34:05',
-    updatedAt: '2025-05-19T09:34:49',
-    documentsCount: 2,
-    eventsCount: 0,
-  },
-  {
-    id: '44c6cd75-b7c4-4e27-b643-ab14c15ee3a1',
-    title: '知识产权案例（已更新）',
-    caseName: 'Ahmed Hassan2',
-    caseType: 'Skilled Worker Visa',
-    description: '这是一个更新后的知识产权侵权案例描述',
-    profileId: 'profile-123',
-    clientId: null,
-    status: CaseStatusEnum.AUTO_FILLING,
-    startDate: null,
-    endDate: null,
-    clientName: null,
-    profileName: null,
-    createdAt: '2025-05-19T09:34:05',
-    updatedAt: '2025-05-19T09:34:49',
-    documentsCount: 2,
-    eventsCount: 0,
-  },
-  {
-    id: '44c6cd75-b7c4-4e27-b643-ab14c15ee3a2',
-    title: '知识产权案例（已更新）',
-    caseName: 'Ahmed Hassan3',
-    caseType: 'Skilled Worker Visa',
-    description: '这是一个更新后的知识产权侵权案例描述',
-    profileId: 'profile-123',
-    clientId: null,
-    status: CaseStatusEnum.HOLD,
-    startDate: null,
-    endDate: null,
-    clientName: null,
-    profileName: null,
-    createdAt: '2025-05-19T09:34:05',
-    updatedAt: '2025-05-19T09:34:49',
-    documentsCount: 2,
-    eventsCount: 0,
-  },
-  {
-    id: '44c6cd75-b7c4-4e27-b643-ab14c15ee3a3',
-    title: '知识产权案例（已更新）',
-    caseName: 'Ahmed Hassan4',
-    caseType: 'Skilled Worker Visa',
-    description: '这是一个更新后的知识产权侵权案例描述',
-    profileId: 'profile-123',
-    clientId: null,
-    status: CaseStatusEnum.READY,
     startDate: null,
     endDate: null,
     clientName: null,
